enable m4 multiply algorithm; fix tiny bug in next_prime that was just introduced minutes ago.

+++ src/sage/rings/integer_ring.pyx
@@ -73,7 +73,7 @@ from random import randrange
 cdef extern from "stdlib.h":
     long random()
     void srandom(unsigned int seed)
-k = randrange(0,Integer(2)**Integer(32))
+k = randrange(0,int(2)**int(32))
 srandom(k)
 
 cdef gmp_randstate_t state
